+++ src/endpoints/explorer/unlock.js
@@ -1,4 +1,5 @@
 import express from 'express';
+import Wayback from 'wayback.js';
 import { Config } from '../../utils/config.js';
 import * as DB from '../../utils/db.js';
 import * as Util from '../../utils/helper.js';
@@ -6,6 +7,9 @@ import * as Web from '../../utils/web.js';
 
 const useRelay = Config.app.relay.server;
 const useWebSource = Config.app.flag.enable_web_source;
+
+const wayback = new Wayback();
+const maxFileAgeInDays = 90;
 
 var router = express.Router();
 
@@ -17,16 +21,25 @@ router.get('/', async function (req, res, next) {
 
     const unlock = await DB.getLensUnlock(lensId);
     if (unlock && unlock[0]) {
-        // trigger re-download to catch missing files automatically
-        await Util.downloadUnlock(unlock[0].lens_id, unlock[0].lens_url);
+        if (unlock[0].lens_id && unlock[0].lens_url) {
+            // trigger re-download to catch missing files automatically
+            await Util.downloadUnlock(unlock[0].lens_id, unlock[0].lens_url);
 
-        return res.json(Util.modifyResponseURLs(unlock[0]));
-    } else {
-        const remoteUnlock = await getRemoteUnlockByLensId(lensId);
-        if (remoteUnlock) {
-            return res.json(remoteUnlock);
+            wayback.saveOutdatedUrl(unlock[0].lens_url, maxFileAgeInDays);
+
+            return res.json(Util.modifyResponseURLs(unlock[0]));
+        } else {
+            console.warn('Unlock Download URL is missing', lensId);
         }
     }
+
+    const remoteUnlock = await getRemoteUnlockByLensId(lensId);
+    if (remoteUnlock) {
+        return res.json(remoteUnlock);
+    }
+
+    // TODO FIX in upcomming versions
+    console.warn('This Lens can not be used. The issue is known and will most likely be fixed in upcomming versions of Snap Camera Server.', lensId);
 
     return res.json({});
 });
@@ -36,7 +49,7 @@ async function getRemoteUnlockByLensId(lensId) {
         if (useRelay) {
             const unlock = await Util.getUnlockFromRelay(lensId);
             if (unlock) {
-                DB.insertUnlock(unlock);
+                await DB.insertUnlock(unlock);
                 return unlock;
             }
         }